backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-js-asset==3.0.1
django-modeltranslation==0.19.11
django-mptt==0.16.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
phonenumbers==8.13.52
pillow==11.0.0
psycopg==3.2.4
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
